filename change dialog now in separate file; requested changes made.
